Moves CronusContext to a separate file
